Perform verification in step authenticate
